common: dims should be multiple of block sizes
This commit fixes #71 and should've been done from the very beginning of
mkl-dnn. So far blocking structure doesn't support dimensions that are
not multiple of corresponding block size. In future we want to fix that
by having some kind of tail processing but have no good ideas how to do
that.