cpu: jit: do not use rdi/rcx registers directly
Since both Linux (system V AMD64 ABI) and Microsoft x64 calling
conventions must be supported we should be really careful with %rdi/%rcx
registers which are used for the first parameter pass. In particular,
abi_param1 and abi_not_param1 should be used so that %rdi/%rcx are
interchanged on different systems.
This patch also fixes #138.