jit: reorder: fix support of unaligned addresses for the sum postop
The non VEX instruction do not allow unaligned addresses as
parameters. We here fix this by adding a movups before the
addps. Because VEX instructions do not have this issue, we now
conditionally use addps/vaddps in order to save a register on AVX
machines.

We also add a test case (sum) that raised the issue initially.

Change-Id: Ib83dde0f8a5f5e2af8a18366bfa4bca7c41a9a70