build: allow using compiler's OpenMP RT
Pass `-DMKLDNN_THREADING=OMP:COMP` to cmake for that.
This is one of the steps required to address #312 issue.

By chance, this fixes #313.
The reason: we used ?undocumented? cmake flag [1] to prevent linking
with compilers OMP RT. Apparently this flag stopped working for recent
cmake versions. The new workaround is: `-Wl,--as-needed` for GCC and
`/full/path/to/libiomp5.so` for clang. How tricky... :(

[1] CMAKE_CXX_CREATE_SHARED_LIBRARY_FORBIDDEN_FLAGS

f: build: allow using compiler's OpenMP RT