benchdnn: rnn: disable small lstm bwd tests
Those need to be disable as benchdnn ref backward implementation assumes that
the forward pass returns the same output in tested impl and ref impl.
However, with the introduction of jitted eltwise in MKL-DNN, we no
more rely on the compiler libm so the forward passes now differ.

We will have to fix our backward testing in benchdnn before enabling
these tests again.