src: cpu: rnn: split temporary gates in scratchpad
We split the temporary gates from workspace to scratchpad.
This has two purposes: 1) the temporary gates can have a different
data type than the workspace, 2) when doing inference, we reduce the
memory footprint.

we also take advantage of that patch to "templatize" many functions to
make them reusable for different datatypes, especially bfloat16.