sycl: remove memory storage check
Primitives may create memory storages internally (for scratchpad memory)
and have no prior knowledge about whether user provided memory
objects are buffer- or USM-based. This is the reason why it's not
possible to implement this check reliably and the library has to support
mixed execution: when buffer/USM memory objects can be mixed.