cpu: common: fix crash when a primitive using a global scratchpad is initialized and executed on different threads
Using thread-local for the global scratchpad is unnecessary and even
buggy. All threads actually share the same scratchpad, which is
created and queried only on the main thread. If the scratchpad is
queried on some thread other than the one it was created on (e.g. the
application calls the API from multiple threads), thread-local causes
a segfault because the scratchpad is uninitialized on the current
thread.