cpu: xbyak: never call exit() on failure
A library should *never* call exit(). If a runtime error occurs, it
should return an error status through the API to let the application
decide how to handle the error. Some Linux distributions (e.g.
Fedora) strongly discourage calling exit() in libraries and they
automatically check this when creating packages.

In this case, it seems that exit() can be triggered only by an
internal error, so it's safe to just remove the exit() calls (or
perhaps could be replaced with assert()).