cpu, ocl: detect and properly handle memory storage allocation and OCL runtime errors
The library checks for small memory allocation errors but it does
*not* report out of memory errors when allocating the *actual* memory
storage. On the CPU these allocations are not checked at all and the
whole application crashes, and on the GPU exit() is called, which also
kills the application. This is a major issue because out of memory
errors usually occur for large allocations, which are not handled
properly (or at all). Also, exit() should never be called inside the
library, and assert() should not be used for checking whether an
allocation was successful. This commit fixes all these issues.