benchdnn: rnn: int8: change validation criteria
This patch makes int8 validation generally stricter.
However, it partially relaxes the condition on `dst_c_last_iter`
(see the comment in the patch).

Hopefully, this commit will be improved later with switching to another
testing method, that involves modulo-operator instead of unreliable playing
with the scaling factors in rnn_tparams.