fixup: examples: use Release build by default
Generally, normal CMake variables hide cache variables. However, in
cases like:

    set(A "1")
    set(A "2" CACHE STRING "")

variable 'A' gets the value '2' given that it was not in the cache
before the first set.

CMAKE_BUILD_TYPE is a cache variable and CMake may set it explicitly via
set(CMAKE_BUILD_TYPE ... CACHE ...) and overwrite the initial value. To
prevent that we should set it as a cache variable. In this case all
subsequent set(CMAKE_BUILD_TYPE ... CACHE .. ) would not have any effect
(unless called with FORCE).