tests: do not explicitly require bash for test_c_symbols (#631)
This fixes up the test script and cmake call to just use the shell, rather than
requiring /bin/bash explicitly, and removes some bashisms. This is relevant for
packaging on distributions/containers that may not have this layout. A run of
the script produces the same output as before.

original commit message:
tests/generate_c-symbols_refs.sh: do not require /bin/bash explicitly