sycl: avoid name collision to workaround an issue
If cmake finds two different directories with two files which have
the same name it creates subdirectories for object files to separate
them. When building DNNL with msbuild the path to such directories contains
wrong slashes which causes build issues. It's most likely an issue with
integration of the DPCPP compiler into Visual Studio.