cpu: conv: x64: Implement bwd_w algorithm targeting nxc data formats
This implements bwd_w algorithm in C++ using a loop order targeting the nxc data
format. This is currently implemented as a microkernel to improve development
speed as there may be some more optimization features that should be added such
as blocking on ic/oc for large ic/oc and threading on groups and possibly on the
kh/kw dimensions. After this implementation is feature complete, it may be
worthwhile to move more loops into the kernel to decrease DNNL's dependency on
an optimizing compiler as the performance can vary by around 10% from minor
changes to the C++ driver code.

The compute portion of this kernel unrolls on ow to amortize updates to the
weights buffers and unrolls on kw and ic to maximize the amount of computation
relative to loads from the src buffer.