src: cpu: x64: eltwise_injector: make table registration more robust
In the previous version of the code, table entries are injected in
prepare_table following the order of the iterator on entry_map_.
However, the offsets are computing wrt the insertion order and we have
no guarantee the two match (they happen to as we insert in key order,
and iterator iterates in key order).

The logic is changed so that we compute the offsets using the
entry_map_ iterator order (the assumption being that iterating is
deterministic).

We also add the capability to control if a constant should be
broadcasted to vlen or not.