include: add max value to dnnl_primtive_kind_t and dnnl_query_t enum
Add a max enum value to prevent UB associated with casts required internal to
dnnl when implementing internal only enum values. The value 0x7fff was chosen to
be portable within the C++ standard. While a larger value should not be an issue
as we are unlikely to support architectures with 2 byte int, it is also unlikely
that over 16,000 enum values will be required.