cpu: x64: jit: remove default operand from uni instructions
Some "universal" instructions in jit_generator allow a default operand
as the second source operand. For vex-encoded instructions, Xbyak
gracefully handles such cases by duplicating the destination register.
For example:
   vaddps(xmm1, op1, operand()) -> vaddps(xmm1, xmm1, op1)

However, the "universal" instructions in jit_generator are also intended
to work for non-vex instruction sets (namely, sse4.1), and a common use
case is something more like:
   uni_vaddps(xmm1, xmm1, op2) -> addps(xmm1, op2)

This can lead to awkward scenario where the second operand is not passed
to the "universal" instruction, so that op2 is an empty operand. To
avoid such pitfalls, it is best to simply remove default operands from
all "universal" instructions.