benchdnn: rnn: align tensor filling for different compilers
Different compilers use different standard library implementations. It
just so happens that for std::normal_distribution, different values are
returned (or the same values in a slightly different order). This leads
to different tensor filling in RNN testing, and increases the chances of
false positive errors being seen in routine testing. This commit adds a
simple implementation for normal_distribution based on the so-called
Marsaglia polar method.