version.cmake: tweak git log invocation
If a user has set log.showSignature = true, then build will break
because DNNL_VERSION_HASH define will be overrun by muliple lines about
the signature of the commit.

If a user has customized core.abbrev-- e.g. Linux kernel developers are
encouraged to set it to 12-- then each build could display a commit hash
string of differing lengths. The git default is "auto", so length of the
hash string could change depending on, e.g., the number of commits in
this repository. That's probably undesirable. By explicitly overriding
abbrev on the command line, it ensures all builds going forward will
generate a hash of the same length (namely, the entire commit hash).

With --oneline one and only one line of output ought to appear, and it
will be the commit hash.

Signed-off-by: Joe Konno <joe.konno@intel.com>

v2: add support for git without support for --no-show-signature
Requested-by: Igor Safonov <igor.safonov@intel.com>