gpu: reorder: refactor kernel into 2 parts, reference & custom (#1660)
The 'simple' reorder is no longer simple, with 19 separate
implementations all placed in a single source file. This change
splits reference kernel into 2 parts: reference (1 implementation
that is slow but supports every legal reorder case) and custom
(remaining implementations which are fast but specialized for
certain, often very limited, sets of cases).
This refactoring is a preparation step before introduction of 20th
reorder implementation which will be fast and generic but very
complicated.