cpu: x64: gemm: reduce nocopy heap/stack usage
If we are over stack capacity we only need to allocate the maximum k
that is called in the leaf sgemm kernel. Largest k value is the cache
blocking size for k-dimension.

If buffer requirement is large page fault will happen in the kernel when
using heap memory.  Using smaller buffers and allocating in each thread
diminishes effect, making malloc happier, however, it is not guaranteed.