cpu: remove all std::map global-statics in favor of local-statics
A global static must be initialized at load-time of the library, which
incurs a performance penalty and emits extra code. Instead, use a
function-local static variable, which gets initialized on-demand (on
first use) by whatever calls it.

This code could be refactored not to use a std::map and instead do a
linear or a binary search on an array. That will have better cache
locality and can be initialized with constexpr.