src: cpu: x64: replace const-ref uses of Xbyak::Operand with regular values
Xbyak::Operand and its derived classes (like Reg, Reg64, Xmm, Ymm, Zmm,
Opmask, etc.) are trivially copyable. They have a weird bit-field
arrangement adding up to 38 bits, so they actually occupy 64 bits in
memory, which is the same as a pointer. Therefore, this change does not
increase memory usage of the classes touched.

However, it improves codegen because no longer are those fields pointers
requiring dereferencing when used with Xbyak itself, nor RIP-relative
addressing when created in the class.

In fact, those members ought to be static constexpr, which would
completely remove them from memory use. However, that's not possible why
oneDNN requires C++11 support. Massive improvement left as future work.

This is not meant to be an exhaustive update. It's only there for the
cases I'm going to need for the next change (and other things that were
in the same file and caught by the regexp).