common: bnorm: avoid assert for fused relu via flags
It would be nice if bnorm pd methods could abstract the particulars of
where the relu postop comes from. To this end, it makes sense to allow
query of alpha value for any form of relu postop. At the same time, it
is good to keep an assert to warn caller of this method that alpha has
no meaning in absence of relu postop.