tests: benchdnn: aarch64: set relative error in conv tests to eps of fp16
The previous value for relative error was zero, which caused
conv with relu to fail after adding support for fp16 on aarch64.
The reason for the failure is that the reference implementation
does the calculations in fp32 and then casts to fp16, instead of
doing the calculations in fp16.
Note that eps for fp16 = 0.000977 ~ 0.001