common: refactor primitive cache structures
Moves most of the primitive_cache_t and lru_primtive_cache_t implementation into
a templated and reusable cache_t and lru_cache_t structures. This required
creating a new `get_or_create()` interface to completely hide concurrency
management within the cache structures. As part of this, the locking logic was
refactored to consistently use the RAII idiom, but most other code was left as
is.