src: cpu: replace full capture by value with analogues
This closes #1249
This closes #1563

C++20 for new GCC compilers report full capture by value as deprecated.
More info may be found here:
https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2018/p0806r2.html

A straightforward replace of [=] with [=, this] will fail for C++17,
that's why replacement used either full capture by reference or specific
variables and objects used in the scope.