x64: jit_io_helper: Add declarations of explicit specializations and instantiations
Currently, code that uses the specializations in question will
accidentally and erroneously specialize the primary template. The
reason this appears to "work" is a pure accident due to the way the
linker currently works. The code violates the one-definition-rule,
and all programs depending on it are ill-formed, no diagnostic
required (i.e. running all such programs has undefined behaviour).

Additionally, declaring explicit instantiations is also good practice,
since it declares intent and prevents accidental later explicit
specializations (it would be an error to declare those after the
extern declaration of the explicit instantiation had been seen).