benchdnn: utils: compare: add a new corr criterion for binary post-op
The problem with binary po is it may lead to catastrophic cancellation,
and there's no way to control the behavior as filling is independent
from the operation. But even when it can be controlled, it isn't always
clear how to make binary po avoid such errors for all supported
driver problems. It's simpler to let some wider range of errors to let
go when binary po presents under assumption that the core operation has
tighter threshold to catch potential issues.