gtest: sycl: Fix error when running SYCL tests with OpenMP CPU runtime
The test failed when the SYCL tests were running with the DNNL_CPU_RUNTIME was
set to OMP and there was a CPU SYCL device available on the system. In this
sceanario the get_count(engine::kind::cpu) returned 1 because of the CPU SYCL
device and the test was not skipped. When the sycl_interop::get_device funciton
tried to create a device object targeting the CPU, an exception was thrown.