gpu: intel: jit: pass: fix to skip cse_exprs only on critical path
The previous algorithm did not consider the allocation stack when deciding which
`let_t` statements to skip. The result is that skipping expressions may not
reduce register consumption. Beyond the unnecessary work this can create, this
also can cause kernel generation failure due to the mis-estimation of GRF usage.

This patch modifies the existing algorithm to run in allocation stack aware
manner and avoid unneeded skips.