Backport cpu: aarch64: matmul: Optimize (A^T)*(B^T) in acl_matmul
This PR computes (B*A)^T instead of (A^T)*(B^T) when the
cost of transposing (B*A) is cheaper. This improves performance
by ~1.25x for square matrices and even higher for
tall-skinny/fat-short matrices.

It also reduces code duplication and moves allocation of
dst accumulator from ACL to scratchpad memory in oneDNN

Co-authored-by: Annop Wongwathanarat <annop.wongwathanarat@arm.com>