Backport cpu: aarch64: matmul: fuse sum post op in acl matmul
Fuse the sum post op in acl matmul by setting the accumulate flag to
true in arm_compute::GEMMInfo. This speeds up the post op and saves
allocating a temporary dst sized tensor.

We also added `_for_sum` to `use_dst_acc` flag to stop it being confused
with the `dst_acc` used for transposing.

Change the way we deal with fused eltwise (as well as the new sum) to
fix segfaults when binary ops followed fused ops.

Co-authored-by: Milos Puzovic <milos.puzovic@arm.com>
Co-authored-by: Jonathan Deakin <jonathan.deakin@arm.com>