xe: introduce dim_idx_t
Introduce a standard type to represent dimension indices. This is intended to
replace the 4 different standards we currently have, such as using dim_id_t,
int, dim_t, and size_t. The type is switched to unsigned to remove
signed/unsigned conversions when using this type for indexing. Using a strong
type like dim_id_t is discarded for now to enable better compatibility with
existing code.