verbose converter: pass ambiguous cases
When runtime dimensions resolve to 1s in matmul cases with bias, the
bits for those dimensions will not be set in the reported bias mask in
the verbose line. In these cases, the implementations for the reference
and computed runs will differ, but the rest of the verbose line will be
the same. Under these circumstances, we allow the test to pass.