cpu: x64: brgemm: add a method to identify when D will be used
This fixes #2179

Reuse the method in ukernel API to tip users when one or another execute
method should be used for a given brgemm configuration. It's impossible
to return an error at creation stage as it depends on the executioin
function whether it suits for a given brgemm object or not.