xe: ukernel: Delay microkernel check until micro_sdpa creation
Checks for the compatibility to build/run microkernel codes were performed at
program initialization time. This commit delays the checks to a later point so
that they are only performed once the micro_sdpa kernel is generated. The reason
for this is that the vISA compiler generates a kernel.errors.txt if there is an
error and this caused the CI to fail because the build directory is dirty
afterwards.