xe: jit: gemm: partial revert of 6b1398
Many of the modified strategies are on architectures that are not under active
development. The fact that these strategies were changed is likely an issue with
the upstream repo. In particular, one of the XeLP modifications triggered an out
of registers issue on the following workload:

```
benchdnn --matmul --engine=gpu --dt=bf16:s4:bf16
--attr-scales=wei:per_ocic:bf16:32x1 --attr-zero-points=wei:common:1:u8
--attr-fpmath=bf16:true 3x96x96:3x96x64
```