xe: ocl: remove unused post-op code
The use of a switch statement in for post-op dispatching induces implementation
complexity. In particular, it requires setting up a number of unused arguments
so that the code compiles and makes the source code significantly more complex.
This PR removes that switch statement and the associated arguments.

Additionally, this PR removes some dead code that became obvious during the refactoring.